big changes
revamped the design and GUI of the whole thing plus the process

some issues:
weird problem with writing data to file where the data for each latest trial would overwrite the data for all previous existing trials?
i.e. instead of progressing from:
a -> a, b -> a, b, c
it would go from:
a -> b, b -> c, c, c

probably something weird to do with putting the whole thing in a while loop??